Extract native sidebar
