+++ Sources/QuillCodeApp/WorkspaceSwiftUIView.swift
@@ -1027,414 +1027,6 @@ private struct QuillCodeSearchView: View {
     }
 }
 
-private struct QuillCodeSidebarView: View {
-    var projects: ProjectListSurface
-    var sidebar: SidebarSurface
-    var commands: [WorkspaceCommandSurface]
-    var onSelectProject: (UUID?) -> Void
-    var onAddProjectRequested: () -> Void
-    var onProjectAction: (ProjectItemActionSurface) -> Void
-    var onSelectThread: (UUID) -> Void
-    var onThreadAction: (SidebarItemActionSurface) -> Void
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            QuillCodeSidebarActionsView(commands: commands, onCommand: onCommand)
-            Divider()
-            HStack {
-                Text(sidebar.title.uppercased())
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                Spacer()
-                if let action = sidebar.bulkActions.first(where: {
-                    sidebar.isSelectionMode ? $0.kind == .clearSelection : $0.kind == .select
-                }) {
-                    Button(action.title) {
-                        onCommand(command(for: action))
-                    }
-                    .font(.caption.weight(.semibold))
-                    .buttonStyle(.plain)
-                    .foregroundStyle(action.isEnabled ? QuillCodePalette.blue : QuillCodePalette.muted)
-                    .disabled(!action.isEnabled)
-                }
-            }
-            if sidebar.isSelectionMode {
-                QuillCodeSidebarBulkActionsView(
-                    selectionLabel: sidebar.selectionLabel,
-                    actions: sidebar.bulkActions.filter { $0.kind != .clearSelection },
-                    onCommand: onCommand
-                )
-            }
-            if sidebar.items.isEmpty {
-                Text(sidebar.emptyTitle)
-                    .font(.callout)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            } else {
-                ScrollView {
-                    LazyVStack(spacing: 8) {
-                        if !sidebar.pinnedItems.isEmpty {
-                            QuillCodeSidebarThreadSectionView(
-                                title: "Pinned",
-                                items: sidebar.pinnedItems,
-                                isSelectionMode: sidebar.isSelectionMode,
-                                onSelectThread: onSelectThread,
-                                onThreadAction: onThreadAction,
-                                onCommand: onCommand
-                            )
-                        }
-                        if !sidebar.recentItems.isEmpty {
-                            QuillCodeSidebarThreadSectionView(
-                                title: "Recent",
-                                items: sidebar.recentItems,
-                                isSelectionMode: sidebar.isSelectionMode,
-                                onSelectThread: onSelectThread,
-                                onThreadAction: onThreadAction,
-                                onCommand: onCommand
-                            )
-                        }
-                        if !sidebar.archivedItems.isEmpty {
-                            QuillCodeSidebarThreadSectionView(
-                                title: "Archived",
-                                items: sidebar.archivedItems,
-                                isSelectionMode: sidebar.isSelectionMode,
-                                onSelectThread: onSelectThread,
-                                onThreadAction: onThreadAction,
-                                onCommand: onCommand
-                            )
-                        }
-                    }
-                }
-            }
-            Divider()
-            QuillCodeProjectListView(
-                projects: projects,
-                onSelectProject: onSelectProject,
-                onAddProjectRequested: onAddProjectRequested,
-                onProjectAction: onProjectAction
-            )
-            Spacer(minLength: 0)
-            QuillCodeSidebarUtilityActionsView(commands: commands, onCommand: onCommand)
-        }
-        .padding(14)
-        .background(QuillCodePalette.sidebar)
-    }
-
-    private func command(for action: SidebarBulkActionSurface) -> WorkspaceCommandSurface {
-        WorkspaceCommandSurface(
-            id: action.commandID,
-            title: action.title,
-            category: WorkspaceCommandPalette.threadCategory,
-            isEnabled: action.isEnabled
-        )
-    }
-}
-
-private struct QuillCodeSidebarBulkActionsView: View {
-    var selectionLabel: String
-    var actions: [SidebarBulkActionSurface]
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text(selectionLabel)
-                .font(.caption)
-                .foregroundStyle(QuillCodePalette.muted)
-            LazyVGrid(columns: [GridItem(.adaptive(minimum: 72), spacing: 6)], spacing: 6) {
-                ForEach(actions) { action in
-                    Button(action.title) {
-                        onCommand(command(for: action))
-                    }
-                    .font(.caption.weight(.semibold))
-                    .padding(.vertical, 6)
-                    .padding(.horizontal, 8)
-                    .frame(maxWidth: .infinity, minHeight: QuillCodeMetrics.minimumHitTarget)
-                    .background((action.isDestructive ? QuillCodePalette.red : QuillCodePalette.panel).opacity(action.isEnabled ? 1 : 0.45))
-                    .foregroundStyle(action.isDestructive ? Color.white : QuillCodePalette.text)
-                    .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-                    .disabled(!action.isEnabled)
-                    .buttonStyle(.plain)
-                }
-            }
-        }
-        .padding(8)
-        .background(QuillCodePalette.background.opacity(0.55))
-        .clipShape(RoundedRectangle(cornerRadius: 16, style: .continuous))
-    }
-
-    private func command(for action: SidebarBulkActionSurface) -> WorkspaceCommandSurface {
-        WorkspaceCommandSurface(
-            id: action.commandID,
-            title: action.title,
-            category: WorkspaceCommandPalette.threadCategory,
-            isEnabled: action.isEnabled
-        )
-    }
-}
-
-private struct QuillCodeSidebarThreadSectionView: View {
-    var title: String
-    var items: [SidebarItemSurface]
-    var isSelectionMode: Bool
-    var onSelectThread: (UUID) -> Void
-    var onThreadAction: (SidebarItemActionSurface) -> Void
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text(title.uppercased())
-                .font(.caption2.weight(.semibold))
-                .foregroundStyle(QuillCodePalette.muted)
-                .padding(.top, 4)
-            ForEach(items) { item in
-                QuillCodeSidebarThreadRowView(
-                    item: item,
-                    isSelectionMode: isSelectionMode,
-                    onSelectThread: onSelectThread,
-                    onThreadAction: onThreadAction,
-                    onCommand: onCommand
-                )
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-private struct QuillCodeSidebarThreadRowView: View {
-    var item: SidebarItemSurface
-    var isSelectionMode: Bool
-    var onSelectThread: (UUID) -> Void
-    var onThreadAction: (SidebarItemActionSurface) -> Void
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        HStack(spacing: 6) {
-            if isSelectionMode {
-                Button {
-                    onCommand(WorkspaceCommandSurface(
-                        id: "thread-selection-toggle:\(item.id.uuidString)",
-                        title: item.isBulkSelected ? "Deselect chat" : "Select chat",
-                        category: WorkspaceCommandPalette.threadCategory
-                    ))
-                } label: {
-                    Image(systemName: item.isBulkSelected ? "checkmark.circle.fill" : "circle")
-                        .font(.system(size: 16, weight: .semibold))
-                        .foregroundStyle(item.isBulkSelected ? QuillCodePalette.blue : QuillCodePalette.muted)
-                        .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                        .contentShape(Rectangle())
-                }
-                .buttonStyle(.plain)
-                .accessibilityLabel(item.isBulkSelected ? "Deselect \(item.title)" : "Select \(item.title)")
-            }
-            Button {
-                if isSelectionMode {
-                    onCommand(WorkspaceCommandSurface(
-                        id: "thread-selection-toggle:\(item.id.uuidString)",
-                        title: item.isBulkSelected ? "Deselect chat" : "Select chat",
-                        category: WorkspaceCommandPalette.threadCategory
-                    ))
-                } else {
-                    onSelectThread(item.id)
-                }
-            } label: {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(item.title)
-                        .font(.callout.weight(.semibold))
-                        .lineLimit(1)
-                    Text(item.subtitle)
-                        .font(.caption)
-                        .foregroundStyle(QuillCodePalette.muted)
-                        .lineLimit(1)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .buttonStyle(.plain)
-            Menu {
-                ForEach(item.actions) { action in
-                    Button(role: action.kind == .delete ? .destructive : nil) {
-                        onThreadAction(action)
-                    } label: {
-                        Text(action.kind.title)
-                    }
-                }
-            } label: {
-                Image(systemName: "ellipsis")
-                    .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .contentShape(Rectangle())
-            }
-            .buttonStyle(.plain)
-        }
-        .padding(10)
-        .background(item.isSelected ? QuillCodePalette.selection : Color.clear)
-        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-    }
-}
-
-private struct QuillCodeSidebarActionsView: View {
-    var commands: [WorkspaceCommandSurface]
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    private var visibleCommands: [WorkspaceCommandSurface] {
-        ["new-chat", "search", "toggle-extensions", "toggle-automations"].compactMap { id in
-            commands.first { $0.id == id }
-        }
-    }
-
-    var body: some View {
-        VStack(spacing: 8) {
-            ForEach(visibleCommands) { command in
-                Button {
-                    onCommand(command)
-                } label: {
-                    Label(displayTitle(for: command), systemImage: systemImage(for: command.id))
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(.vertical, 8)
-                        .padding(.horizontal, 10)
-                        .background(command.id == "new-chat" ? QuillCodePalette.panel.opacity(0.74) : Color.clear)
-                        .clipShape(RoundedRectangle(cornerRadius: 9, style: .continuous))
-                }
-                .buttonStyle(.plain)
-                .disabled(!command.isEnabled)
-            }
-        }
-    }
-
-    private func displayTitle(for command: WorkspaceCommandSurface) -> String {
-        switch command.id {
-        case "toggle-extensions":
-            return "Plugins"
-        default:
-            return command.title
-        }
-    }
-
-    private func systemImage(for commandID: String) -> String {
-        switch commandID {
-        case "new-chat":
-            return "square.and.pencil"
-        case "search":
-            return "magnifyingglass"
-        case "toggle-automations":
-            return "clock.arrow.circlepath"
-        case "toggle-extensions":
-            return "puzzlepiece.extension"
-        default:
-            return "circle"
-        }
-    }
-}
-
-private struct QuillCodeSidebarUtilityActionsView: View {
-    var commands: [WorkspaceCommandSurface]
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    private var visibleCommands: [WorkspaceCommandSurface] {
-        [
-            "toggle-terminal",
-            "toggle-browser",
-            "toggle-memories",
-            "toggle-activity",
-            "command-palette"
-        ].compactMap { id in
-            commands.first { $0.id == id }
-        }
-    }
-
-    private var settingsCommand: WorkspaceCommandSurface? {
-        commands.first { $0.id == "settings" }
-    }
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Menu {
-                ForEach(visibleCommands) { command in
-                    Button {
-                        onCommand(command)
-                    } label: {
-                        Label(displayTitle(for: command), systemImage: systemImage(for: command.id))
-                    }
-                    .disabled(!command.isEnabled)
-                }
-            } label: {
-                Label("Tools", systemImage: "wrench.and.screwdriver")
-                    .font(.callout.weight(.semibold))
-                    .frame(maxWidth: .infinity, minHeight: QuillCodeMetrics.minimumHitTarget)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .background(QuillCodePalette.panel.opacity(0.50))
-                    .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-            }
-            .buttonStyle(.plain)
-            .help("Tools")
-
-            if let settingsCommand {
-                Button {
-                    onCommand(settingsCommand)
-                } label: {
-                    Image(systemName: systemImage(for: settingsCommand.id))
-                        .font(.system(size: 15, weight: .semibold))
-                        .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                        .foregroundStyle(settingsCommand.isEnabled ? QuillCodePalette.muted : QuillCodePalette.muted.opacity(0.45))
-                        .background(QuillCodePalette.panel.opacity(0.50))
-                        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                }
-                .buttonStyle(.plain)
-                .disabled(!settingsCommand.isEnabled)
-                .help(displayTitle(for: settingsCommand))
-                .accessibilityLabel(displayTitle(for: settingsCommand))
-            }
-        }
-        .padding(.top, 10)
-    }
-
-    private func displayTitle(for command: WorkspaceCommandSurface) -> String {
-        switch command.id {
-        case "toggle-extensions":
-            return "Plugins"
-        case "toggle-terminal":
-            return "Terminal"
-        case "toggle-browser":
-            return "Browser"
-        case "toggle-memories":
-            return "Memories"
-        case "toggle-activity":
-            return "Activity"
-        case "command-palette":
-            return "Command palette"
-        case "settings":
-            return "Settings"
-        default:
-            return command.title
-        }
-    }
-
-    private func systemImage(for commandID: String) -> String {
-        switch commandID {
-        case "toggle-extensions":
-            return "puzzlepiece.extension"
-        case "toggle-automations":
-            return "clock.arrow.circlepath"
-        case "toggle-terminal":
-            return "terminal"
-        case "terminal-clear":
-            return "clear"
-        case "toggle-browser":
-            return "globe"
-        case "toggle-memories":
-            return "brain.head.profile"
-        case "toggle-activity":
-            return "waveform.path.ecg"
-        case "command-palette":
-            return "command"
-        case "settings":
-            return "gearshape"
-        default:
-            return "circle"
-        }
-    }
-}
-
 private struct QuillCodeWorktreeCreateView: View {
     @Binding var draft: QuillCodeWorktreeCreateDraft
     var onCancel: () -> Void
@@ -1542,101 +1134,6 @@ private struct QuillCodeWorktreeRemoveView: View {
         .padding(24)
         .frame(width: 520)
         .background(QuillCodePalette.background)
-    }
-}
-
-private struct QuillCodeProjectListView: View {
-    var projects: ProjectListSurface
-    var onSelectProject: (UUID?) -> Void
-    var onAddProjectRequested: () -> Void
-    var onProjectAction: (ProjectItemActionSurface) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack {
-                Text(projects.title.uppercased())
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                Spacer()
-                Button(action: onAddProjectRequested) {
-                    Image(systemName: "plus.circle")
-                        .imageScale(.small)
-                        .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                        .contentShape(Rectangle())
-                }
-                .buttonStyle(.plain)
-                .foregroundStyle(QuillCodePalette.muted)
-                .help("Open project")
-                Button {
-                    onSelectProject(nil)
-                } label: {
-                    Image(systemName: "xmark.circle")
-                        .imageScale(.small)
-                        .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                        .contentShape(Rectangle())
-                }
-                .buttonStyle(.plain)
-                .foregroundStyle(QuillCodePalette.muted)
-                .help("Clear project")
-            }
-            if projects.items.isEmpty {
-                Text(projects.emptyTitle)
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-            } else {
-                ForEach(projects.items) { project in
-                    HStack(spacing: 6) {
-                        Button {
-                            onSelectProject(project.id)
-                        } label: {
-                            VStack(alignment: .leading, spacing: 3) {
-                                HStack(spacing: 6) {
-                                    Text(project.name)
-                                        .font(.callout.weight(.semibold))
-                                        .lineLimit(1)
-                                    if project.isRemote {
-                                        Text(project.connectionKindLabel)
-                                            .font(.caption2.weight(.bold))
-                                            .foregroundStyle(QuillCodePalette.blue)
-                                            .padding(.horizontal, 6)
-                                            .padding(.vertical, 2)
-                                            .background(QuillCodePalette.blue.opacity(0.14))
-                                            .clipShape(Capsule())
-                                    }
-                                }
-                                Text(project.path)
-                                    .font(.caption)
-                                    .foregroundStyle(QuillCodePalette.muted)
-                                    .lineLimit(1)
-                            }
-                            .frame(maxWidth: .infinity, alignment: .leading)
-                        }
-                        .buttonStyle(.plain)
-
-                        Menu {
-                            ForEach(project.actions) { action in
-                                Button(role: action.kind == .remove ? .destructive : nil) {
-                                    onProjectAction(action)
-                                } label: {
-                                    Text(action.kind.title)
-                                }
-                                .disabled(!action.isEnabled)
-                                .help(action.disabledReason ?? action.kind.title)
-                            }
-                        } label: {
-                            Image(systemName: "ellipsis")
-                                .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                                .foregroundStyle(QuillCodePalette.muted)
-                                .contentShape(Rectangle())
-                        }
-                        .buttonStyle(.plain)
-                    }
-                    .padding(10)
-                    .background(project.isSelected ? QuillCodePalette.selection : Color.clear)
-                    .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                }
-            }
-        }
     }
 }
